Fix:  Custom wallpapers doesn't mount into wallpaper switching cycle #1136
Closes #1136
