+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Typos,spelling and and cleanup
 - Dunst: Fix dunst crashing when the font cannot handle unsupported characters -- Thanks to [#1131](https://github.com/HyDE-Project/HyDE/issues/1131)
 - UWSM: Clean up the xdg freedesktop.org spec as uwsm handles it
+- Wallpaper: fix #1136 as exporting arrays are not supported in bash
+
 
 ### Changed
